refactor: consolidate useWagmi into useWallet
